feat: storages: registry (artifacts and certs); api (db and uploaded; sso (db)

+++ 00-cluster/00-00-setup-environment/setup-environment.sh
@@ -2,6 +2,10 @@
 
 set -xe
 
+if [[ ! "${LADESA_DEPLOY_SETUP_CLUSTER_ENVIRONMENT}" == "true" ]]; then
+  exit 0;
+fi
+
 TMP_SETUP_PATH=/tmp/cluster-setup
 
 rm -rf ${TMP_SETUP_PATH};
